refactor: move pin label validation from Chip to NormalComponent base class
